refactor: move snippet to taskfile package
