+++ taskfile/reader.go
@@ -193,10 +193,10 @@ func (r *Reader) readNode(node Node) (*ast.Taskfile, error) {
 	var tf ast.Taskfile
 	if err := yaml.Unmarshal(b, &tf); err != nil {
 		// Decode the taskfile and add the file info the any errors
-		taskfileInvalidErr := &errors.TaskfileDecodeError{}
-		if errors.As(err, &taskfileInvalidErr) {
-			snippet := errors.NewSnippet(b, taskfileInvalidErr.Line, taskfileInvalidErr.Column, 2)
-			return nil, taskfileInvalidErr.WithFileInfo(node.Location(), snippet)
+		taskfileDecodeErr := &errors.TaskfileDecodeError{}
+		if errors.As(err, &taskfileDecodeErr) {
+			snippet := NewSnippet(b, taskfileDecodeErr.Line, taskfileDecodeErr.Column, 2)
+			return nil, taskfileDecodeErr.WithFileInfo(node.Location(), snippet.String())
 		}
 		return nil, &errors.TaskfileInvalidError{URI: filepathext.TryAbsToRel(node.Location()), Err: err}
 	}
