+++ taskfile/snippet.go
@@ -1,4 +1,4 @@
-package errors
+package taskfile
 
 import (
 	"bytes"
