+++ taskfile/snippet_test.go
@@ -1,4 +1,4 @@
-package errors
+package taskfile
 
 import (
 	"strings"
